added unit tests

+++ src/main/java/praktikum/Bun.java
@@ -7,9 +7,9 @@ package praktikum;
 public class Bun {
 
     public String name;
-    public float price;
+    public int price;
 
-    public Bun(String name, float price) {
+    public Bun(String name, int price) {
         this.name = name;
         this.price = price;
     }
@@ -18,7 +18,7 @@ public class Bun {
         return name;
     }
 
-    public float getPrice() {
+    public int getPrice() {
         return price;
     }
