+++ src/main/java/praktikum/Ingredient.java
@@ -9,15 +9,15 @@ public class Ingredient {
 
     public IngredientType type;
     public String name;
-    public float price;
+    public int price;
 
-    public Ingredient(IngredientType type, String name, float price) {
+    public Ingredient(IngredientType type, String name, int price) {
         this.type = type;
         this.name = name;
         this.price = price;
     }
 
-    public float getPrice() {
+    public int getPrice() {
         return price;
     }
